Refactor ServiceAccountsDataGrid
