--- Source/SelfService/Web/integrations/connection/connectionDetails/consumeDataEventStreams/serviceAccounts/ServiceAccountsDataGrid.tsx
@@ -1,102 +1,0 @@
-// Copyright (c) Aigonix. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-import React, { useMemo } from 'react';
-
-import { DataGridPro, GridColDef } from '@mui/x-data-grid-pro';
-
-import { Button, DataGridCustomToolbar, dataGridDefaultProps, DataGridWrapper, StatusIndicator } from '@dolittle/design-system';
-
-import { KafkaServiceAccountListDto } from '../../../../../apis/integrations/generated';
-
-import { formatDate } from '../../../../../utils/helpers/dates';
-
-export type ServiceAccountsDataGridProps = {
-    serviceAccountsDataGridRows: KafkaServiceAccountListDto[];
-    isLoading: boolean;
-    onServiceAccountCreate: () => void;
-    onServiceAccountDelete: () => void;
-    onViewCertificate: (serviceAccount: KafkaServiceAccountListDto) => void;
-    onViewKey: (serviceAccount: KafkaServiceAccountListDto) => void;
-    onSelectionChanged: (selection: string[]) => void;
-};
-
-export const ServiceAccountsDataGrid = ({ serviceAccountsDataGridRows, isLoading, onServiceAccountCreate, onServiceAccountDelete, onViewCertificate, onViewKey, onSelectionChanged }: ServiceAccountsDataGridProps) => {
-    const serviceAccountsDataGridColumns: GridColDef<KafkaServiceAccountListDto>[] = useMemo(() => [
-        {
-            field: 'serviceAccountName',
-            headerName: 'Name',
-            minWidth: 120,
-            flex: 1,
-        },
-        {
-            field: 'description',
-            headerName: 'Description',
-            minWidth: 250,
-            flex: 1,
-        },
-        {
-            field: 'certificates',
-            headerName: 'Certificates',
-            minWidth: 300,
-            sortable: false,
-            flex: 1,
-            renderCell: params => (
-                <>
-                    <Button variant='outlined' label='View Certificate' onClick={() => onViewCertificate(params.row)} sx={{ mr: 2 }} />
-                    <Button variant='outlined' label='View Key' onClick={() => onViewKey(params.row)} />
-                </>
-            ),
-        },
-        {
-            field: 'createdAt',
-            headerName: 'Created at',
-            minWidth: 100,
-            flex: 1,
-            valueFormatter: params => params.value ? formatDate(params.value) : '-',
-            renderCell: params => <span title={params.value?.toISOString()}>{params.formattedValue}</span>,
-        },
-        {
-            field: 'certificateExpiry',
-            headerName: 'Certificate Validity',
-            minWidth: 100,
-            flex: 1,
-            renderCell: params => {
-                const isValid = params.value > Date.now();
-
-                return (
-                    <>
-                        {params.value
-                            ? <span title={params.value.toISOString()}>
-                                <StatusIndicator status={isValid ? 'success' : 'error'} label={isValid ? 'Up to date' : 'Expired'} />
-                            </span>
-                            : '-'
-                        }
-                    </>
-                );
-            },
-        },
-    ], [onViewCertificate, onViewKey]);
-
-    return (
-        <DataGridWrapper background='dark'>
-            <DataGridPro
-                {...dataGridDefaultProps}
-                rows={serviceAccountsDataGridRows}
-                columns={serviceAccountsDataGridColumns}
-                loading={isLoading}
-                getRowId={row => row.serviceAccountName}
-                checkboxSelection
-                onSelectionModelChange={model => onSelectionChanged(model as string[])}
-                components={{
-                    Toolbar: () => (
-                        <DataGridCustomToolbar title='Your Service Accounts'>
-                            <Button label='Generate New Service Account' startWithIcon='AddCircle' onClick={onServiceAccountCreate} />
-                            <Button label='Delete Service Accounts' color='error' startWithIcon='DeleteRounded' onClick={onServiceAccountDelete} />
-                        </DataGridCustomToolbar>
-                    ),
-                }}
-            />
-        </DataGridWrapper>
-    );
-};
